% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_win_matrix.R
\name{get_win_matrix}
\alias{get_win_matrix}
\title{Win matrix calculator}
\usage{
get_win_matrix(dice_set, dice_ids)
}
\arguments{
\item{dice_set}{The set of dice to be calculated for (a list of numeric vectors).
There must be at least two dice to compare.}

\item{dice_ids}{The identifiers for each dice, labels for convenience.}
}
\value{
A single numeric decimal, representing probability of first die beating
the second die.
}
\description{
This function takes a list of a set of dice, and outputs a data frame representing
the probabilities of games between the dice. The interpretation of entry (i,j)
is the probability that dice i beats dice j.
}
\examples{
get_win_matrix(list(c(1, 2, 3, 4, 5, 6), c(1, 2, 3, 4, 5, 6)), c("dice_a", "dice_b"))

}
\keyword{matrix}
\keyword{probability}
\keyword{win}
