% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grime_dice.R
\name{grime_dice}
\alias{grime_dice}
\title{Grime dice generator}
\usage{
grime_dice()
}
\value{
List of five numeric vectors, each representing one of the Grime dice.
}
\description{
Generates Dr. James Grime's set of nontransitive dice. It takes no parameters.
It returns a list of five vectors, representing the Grime dice.
}
\examples{
grime_dice()

}
\keyword{generate}
