% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grime_dice_names.R
\name{grime_dice_names}
\alias{grime_dice_names}
\title{Grime dice name generator}
\usage{
grime_dice_names()
}
\value{
Vectors of five strings, each representing one of the Grime dice.
}
\description{
Generates names of Dr. James Grime's set of nontransitive dice. It takes no parameters.
It returns a vector of the five names of Grime dice, in length increasing order.
}
\examples{
grime_dice_names()

}
\keyword{generate}
