% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_win.R
\name{prob_win}
\alias{prob_win}
\title{Win probability calculator}
\usage{
prob_win(dice_a, dice_b)
}
\arguments{
\item{dice_a}{The first dice, which we calculate the probability of winning
against \code{dice_b}.}

\item{dice_b}{The second dice, which we use to calculate the probability that
\code{dice_a} beats.}
}
\value{
A single numeric decimal, representing probability of first die beating
the second die.
}
\description{
This function calculates the probability of one dice rolling a higher digit
than a second dice. This function works for any two dice, even of different
lengths. The probability calculated is the first dice beating the second.
}
\examples{
prob_win(c(1, 2, 3, 4, 5, 6), c(1, 2, 3, 4, 5, 6))

}
\keyword{probability}
